pyproject.toml
[tool.poetry]
name = "langchain-project"
version = "0.1.0"
description = ""
authors = ["yeondu-0 <dpdls1021@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
python-dotenv = "^1.2.1"
langchain-text-splitters = "^1.1.0"
langchain-core = "^1.2.6"
langchain-openai = "^1.1.6"
pymupdf = "^1.26.7"
langchain-upstage = "^0.7.5"
faiss-cpu = "^1.13.2"
unstructured = "^0.18.26"
lxml = "^6.0.2"
sentence-transformers = "^5.2.0"
langchain = "^1.2.3"
langchain-community = "^0.4.1"
langchain-qdrant = "^1.1.0"
langchain-huggingface = "^1.2.0"
qdrant-client = "^1.16.2"
streamlit = "^1.53.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
